Add background behind icons
