--- todo
@@ -1,3 +1,0 @@
-rustgpu
-remove the icon overlay background, use sdf in the original background
-maybe have the unused icons on a second row?
